refactor: Update Monei payment service classes and logger usage
- Replace instances of LoggerInterface with the new Logger class in Capture and Refund command classes to standardize logging practices.
- Enhance the Refund command to include detailed logging of refund reasons, improving traceability and error handling.
- Remove obsolete service classes related to payment operations, streamlining the codebase and reducing complexity.
- Update dependency injection configurations to reflect the new service structure, ensuring proper service resolution.
