+++ Service/Api/CancelPayment.php
@@ -6,15 +6,14 @@
 
 declare(strict_types=1);
 
-namespace Monei\MoneiPayment\Service;
+namespace Monei\MoneiPayment\Service\Api;
 
 use Magento\Framework\Exception\LocalizedException;
 use Monei\Model\CancelPaymentRequest;
 use Monei\Model\Payment;
 use Monei\Model\PaymentCancellationReason;
 use Monei\MoneiPayment\Api\Service\CancelPaymentInterface;
-use Monei\MoneiPayment\Service\Api\ApiExceptionHandler;
-use Monei\MoneiPayment\Service\Api\MoneiApiClient;
+use Monei\MoneiPayment\Service\Logger;
 use Monei\MoneiClient;
 
 /**
