+++ Service/Api/ConfirmPayment.php
@@ -6,7 +6,7 @@
 
 declare(strict_types=1);
 
-namespace Monei\MoneiPayment\Service;
+namespace Monei\MoneiPayment\Service\Api;
 
 use Magento\Framework\Exception\LocalizedException;
 use Monei\Model\ConfirmPaymentRequest;
@@ -15,8 +15,7 @@ use Monei\Model\PaymentBillingDetails;
 use Monei\Model\PaymentCustomer;
 use Monei\Model\PaymentShippingDetails;
 use Monei\MoneiPayment\Api\Service\ConfirmPaymentInterface;
-use Monei\MoneiPayment\Service\Api\ApiExceptionHandler;
-use Monei\MoneiPayment\Service\Api\MoneiApiClient;
+use Monei\MoneiPayment\Service\Logger;
 use Monei\MoneiClient;
 
 /**
