+++ Service/Api/VerifyApplePayDomain.php
@@ -7,14 +7,13 @@
 
 declare(strict_types=1);
 
-namespace Monei\MoneiPayment\Service;
+namespace Monei\MoneiPayment\Service\Api;
 
 use Magento\Framework\Exception\LocalizedException;
 use Monei\Model\ApplePayDomainRegister200Response;
 use Monei\Model\RegisterApplePayDomainRequest;
 use Monei\MoneiPayment\Api\Service\VerifyApplePayDomainInterface;
-use Monei\MoneiPayment\Service\Api\ApiExceptionHandler;
-use Monei\MoneiPayment\Service\Api\MoneiApiClient;
+use Monei\MoneiPayment\Service\Logger;
 use Monei\MoneiClient;
 
 /**
