+++ etc/di.xml
@@ -32,52 +32,56 @@
     type="Monei\MoneiPayment\Model\Config\MoneiPaypalPaymentModuleConfig"
   />
   <preference
-    for="Monei\MoneiPayment\Api\Service\CreatePaymentInterface"
-    type="Monei\MoneiPayment\Service\CreatePayment"
+    for="Monei\MoneiPayment\Api\Service\Checkout\SaveTokenizationInterface"
+    type="Monei\MoneiPayment\Service\Checkout\SaveTokenization"
   />
   <preference
-    for="Monei\MoneiPayment\Api\Service\Checkout\CreateLoggedMoneiPaymentInSiteInterface"
-    type="Monei\MoneiPayment\Service\Checkout\CreateLoggedMoneiPaymentInSite"
+    for="Monei\MoneiPayment\Api\Service\GetPaymentInterface"
+    type="Monei\MoneiPayment\Service\Api\GetPayment"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\GetPaymentMethodsInterface"
+    type="Monei\MoneiPayment\Service\Api\GetPaymentMethods"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\CancelPaymentInterface"
+    type="Monei\MoneiPayment\Service\Api\CancelPayment"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\RefundPaymentInterface"
+    type="Monei\MoneiPayment\Service\Api\RefundPayment"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\CapturePaymentInterface"
+    type="Monei\MoneiPayment\Service\Api\CapturePayment"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\ConfirmPaymentInterface"
+    type="Monei\MoneiPayment\Service\Api\ConfirmPayment"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\CreatePaymentInterface"
+    type="Monei\MoneiPayment\Service\Api\CreatePayment"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\VerifyApplePayDomainInterface"
+    type="Monei\MoneiPayment\Service\Api\VerifyApplePayDomain"
+  />
+  <preference
+    for="Monei\MoneiPayment\Api\Service\GenerateInvoiceInterface"
+    type="Monei\MoneiPayment\Service\GenerateInvoice"
   />
   <preference
     for="Monei\MoneiPayment\Api\Service\Checkout\CreateLoggedMoneiPaymentVaultInterface"
     type="Monei\MoneiPayment\Service\Checkout\CreateLoggedMoneiPaymentVault"
   />
   <preference
+    for="Monei\MoneiPayment\Api\Service\Checkout\CreateLoggedMoneiPaymentInSiteInterface"
+    type="Monei\MoneiPayment\Service\Checkout\CreateLoggedMoneiPaymentInSite"
+  />
+  <preference
     for="Monei\MoneiPayment\Api\Service\Checkout\CreateGuestMoneiPaymentInSiteInterface"
     type="Monei\MoneiPayment\Service\Checkout\CreateGuestMoneiPaymentInSite"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\Checkout\SaveTokenizationInterface"
-    type="Monei\MoneiPayment\Service\Checkout\SaveTokenization"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\GetPaymentInterface"
-    type="Monei\MoneiPayment\Service\GetPayment"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\GetPaymentMethodsInterface"
-    type="Monei\MoneiPayment\Service\GetPaymentMethods"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\CancelPaymentInterface"
-    type="Monei\MoneiPayment\Service\CancelPayment"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\RefundPaymentInterface"
-    type="Monei\MoneiPayment\Service\RefundPayment"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\CapturePaymentInterface"
-    type="Monei\MoneiPayment\Service\CapturePayment"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\ConfirmPaymentInterface"
-    type="Monei\MoneiPayment\Service\ConfirmPayment"
-  />
-  <preference
-    for="Monei\MoneiPayment\Api\Service\GenerateInvoiceInterface"
-    type="Monei\MoneiPayment\Service\GenerateInvoice"
   />
   <preference
     for="Magento\Sales\Controller\Adminhtml\Order\Creditmemo\Save"
@@ -599,7 +603,7 @@
   />
 
   <!-- Type configurations for standardized SDK services -->
-  <type name="Monei\MoneiPayment\Service\CreatePayment">
+  <type name="Monei\MoneiPayment\Service\Api\CreatePayment">
     <arguments>
       <argument name="logger" xsi:type="object">Monei\MoneiPayment\Service\Logger</argument>
       <argument name="exceptionHandler" xsi:type="object"
@@ -672,7 +676,7 @@
         >Monei\MoneiPayment\Api\Config\MoneiPaymentModuleConfigInterface</argument
       >
       <argument name="createPayment" xsi:type="object"
-        >Monei\MoneiPayment\Service\CreatePayment</argument
+        >Monei\MoneiPayment\Api\Service\CreatePaymentInterface</argument
       >
       <argument name="getPaymentService" xsi:type="object"
         >Monei\MoneiPayment\Api\Service\GetPaymentInterface</argument
@@ -703,7 +707,7 @@
         >Monei\MoneiPayment\Api\Config\MoneiPaymentModuleConfigInterface</argument
       >
       <argument name="createPayment" xsi:type="object"
-        >Monei\MoneiPayment\Service\CreatePayment</argument
+        >Monei\MoneiPayment\Api\Service\CreatePaymentInterface</argument
       >
       <argument name="getPaymentService" xsi:type="object"
         >Monei\MoneiPayment\Api\Service\GetPaymentInterface</argument
@@ -734,7 +738,7 @@
         >Magento\Vault\Api\PaymentTokenManagementInterface</argument
       >
       <argument name="createPayment" xsi:type="object"
-        >Monei\MoneiPayment\Service\CreatePayment</argument
+        >Monei\MoneiPayment\Api\Service\CreatePaymentInterface</argument
       >
       <argument name="getPaymentService" xsi:type="object"
         >Monei\MoneiPayment\Api\Service\GetPaymentInterface</argument
@@ -938,7 +942,7 @@
 
   <preference
     for="Monei\MoneiPayment\Api\Service\VerifyApplePayDomainInterface"
-    type="Monei\MoneiPayment\Service\VerifyApplePayDomain"
+    type="Monei\MoneiPayment\Service\Api\VerifyApplePayDomain"
   />
 
   <!-- Configuration for JSON validators -->
@@ -947,6 +951,25 @@
       <argument name="policyInterface" xsi:type="object"
         >Magento\Framework\Authorization\Policy\DefaultPolicy</argument
       >
+    </arguments>
+  </type>
+
+  <!-- Gateway Command Types -->
+  <type name="Monei\MoneiPayment\Gateway\Command\Refund">
+    <arguments>
+      <argument name="refundPaymentService" xsi:type="object"
+        >Monei\MoneiPayment\Api\Service\RefundPaymentInterface</argument
+      >
+      <argument name="logger" xsi:type="object">Monei\MoneiPayment\Service\Logger</argument>
+    </arguments>
+  </type>
+
+  <type name="Monei\MoneiPayment\Gateway\Command\Capture">
+    <arguments>
+      <argument name="capturePaymentService" xsi:type="object"
+        >Monei\MoneiPayment\Api\Service\CapturePaymentInterface</argument
+      >
+      <argument name="logger" xsi:type="object">Monei\MoneiPayment\Service\Logger</argument>
     </arguments>
   </type>
 </config>
